Adding basic view and functionality to OptionsWindow:
- opens after proper login
- has 'disconnect' button
- contains 'msg' label (cut from login window)

NOTE: There's some kind of strange problem with closing client side application (it definitely was there before this and one before check-ins. How to repro: run client app and close it. If opened with vs2012, you will see that it's still debugging. If opened by .exe, open windows manager and you will see that the process is still active. I've tried to kill any threads and ect. in MainWindow Close function but it didn't work (now it's //TO TO). I've submitted the issue for that.